pyproject.toml
[project]
name = "ark-gui-model-test"
version = "0.1.0"
description = "doubao-seed-1-6-vision-250815 GUI model performance test"
requires-python = ">=3.12"
dependencies = [
    "volcengine-python-sdk>=5.0.0",
    "openai>=1.40.0",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
